- Aggiunto messaggio 'dati invalidi' in fase di creazione di nuovo utente - Implementata logica per cancellazione e creazione nuovo utente

+++ src/app/pages/user-list/user-list.component.ts
@@ -5,6 +5,8 @@ import {UserCardComponent} from '../../components/user-list/user-card/user-card.
 import {UserListHeaderComponent} from '../../components/user-list/user-list-header/user-list-header.component';
 import {CreateUserDialogComponent} from '../../components/user-list/create-user-dialog/create-user-dialog.component';
 import {MatDialog} from '@angular/material/dialog';
+import {ConfirmDialogComponent} from '../../components/shared/confirm-dialog/confirm-dialog.component';
+import {MatSnackBar} from '@angular/material/snack-bar';
 
 
 @Component({
@@ -19,13 +21,13 @@ import {MatDialog} from '@angular/material/dialog';
 export class UserListComponent implements OnInit {
 
   private readonly userService = inject(UsersService);
-  newUserDialog = inject(MatDialog);
 
+  dialog = inject(MatDialog);
+  snackBar = inject(MatSnackBar) ;
   users!: User[];
   page: number = 1;
   per_page: number = 10;
   total: number = 0;
-  // TODO: Implementare name e email come variabili di un formgroup
   name!: string;
   email!: string;
   gridCols: number = 3;
@@ -61,18 +63,50 @@ export class UserListComponent implements OnInit {
     this.openCreateUserDialog();
   }
 
-  private openCreateUserDialog(): void {
+  openCreateUserDialog(): void {
     console.log("user-list says: opening dialog");
-    this.newUserDialog.open(CreateUserDialogComponent);
+    this.dialog.open(CreateUserDialogComponent).afterClosed().subscribe(
+      data => {
+        console.log("user-list says: sending data...", data);
+        if(!data) {
+          return;
+        }
+        this.userService.createUser(data).subscribe({
+          next: () => {
+            this.snackBar.open('User created successfully', 'OK', {duration: 3000});
+            this.load();
+            },
+          error: (err) => {
+            const msgError = err?.error?.message || "Creation error";
+            this.snackBar.open(msgError, 'Chiudi', {duration: 3000});
+          }
+        })
+      }
+    );
   }
 
   onDeleteUser(id: number): void {
     console.log("user-list says deleting user with ", id);
-    this.userService.deleteUser(id).subscribe(
-      res => {
-        console.log(res);
-        this.load();
+    const ref = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Confirm Delete User',
+        content: 'Are you sure you want to delete this user?',
       }
-    )
+    });
+    ref.afterClosed().subscribe( ok => {
+      console.log("confirm-dialog says after closed: ", ok);
+      if (!ok) {
+        return
+      }
+      this.userService.deleteUser(id).subscribe({
+        next: () => {
+          this.snackBar.open('User deleted successfully.', 'OK', {duration: 3000});
+          this.load();
+          },
+        error: () => {
+          this.snackBar.open('User deleted failed', 'Chiudi', {duration: 2000});
+        }
+      })
+    })
   }
 }
